+++ src/core/__tests__/model.test.ts
@@ -1,7 +1,29 @@
+import { Model } from '@react3l/react3l/core/model';
+import { Moment } from 'moment';
 import 'reflect-metadata';
-import {TestUser} from '@react3l/react3l/__tests__/models/TestUser';
 
 test('model', () => {
+  class TestClass extends Model {
+    public id?: number;
+
+    public code?: string;
+
+    public name?: string;
+  }
+  class TestUser extends Model {
+    public code?: string;
+
+    public name?: string;
+
+    public email?: string;
+
+    public birthday?: Moment;
+
+    public userClassId?: number;
+
+    public userClass?: TestClass;
+  }
+
   const testId: number = 1;
   const testName: string = 'Test Name';