+++ src/core/order-type.ts
@@ -1,4 +1,6 @@
 /**
  * Order Type
+ *
+ * @deprecated Import OrderType from `core` is deprecated. Import from `types` instead.
  */
 export type OrderType = 'ASC' | 'DESC' | null;