+++ src/form/components/FormField.ts
@@ -8,6 +8,10 @@ export interface FormFieldProps {
   fieldName: string;
 }
 
+/**
+ *
+ * @deprecated FormField is deprecated.
+ */
 export function FormField(props: FormFieldProps): null {
   return null;
 }