+++ src/types/OrderType.ts
@@ -1,0 +1,4 @@
+/**
+ * Order Type
+ */
+export type OrderType = 'ASC' | 'DESC' | null;